# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/memberships": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all published membership tiers sorted by resulting points in ascending order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memberships"
                ],
                "summary": "List all published membership tiers",
                "responses": {
                    "200": {
                        "description": "List of memberships retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Membership"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error or failed to communicate with Directus",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/memberships/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculate points from total payments (100,000 VND = 10 points) and determine tier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memberships"
                ],
                "summary": "Get customer membership info",
                "responses": {
                    "200": {
                        "description": "Customer membership info",
                        "schema": {
                            "$ref": "#/definitions/api.MembershipResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a Stripe payment intent and records the payment information in the database for a booking.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "parameters": [
                    {
                        "description": "Payment creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.CreatePaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error or failed to communicate with Stripe/Directus",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payments/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Confirms a Stripe payment intent and updates the payment record in the database with the confirmation status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Confirm an existing payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment confirmation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ConfirmPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment confirmed successfully",
                        "schema": {
                            "$ref": "#/definitions/db.Payment"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error or failed to confirm payment in Stripe/Directus",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ConfirmPaymentRequest": {
            "type": "object",
            "required": [
                "payment_intent_id",
                "payment_method_id"
            ],
            "properties": {
                "payment_intent_id": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "string"
                }
            }
        },
...
        "api.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "booking_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "booking_id": {
                    "type": "string"
                }
            }
        },
        "api.CreatePaymentResponse": {
            "type": "object",
            "properties": {
                "client_secret": {
                    "type": "string"
                },
                "payment": {
                    "$ref": "#/definitions/db.Payment"
                },
                "publishable_key": {
                    "type": "string"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.MembershipResponse": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "number"
                },
                "early_buy_time": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "tier": {
                    "type": "string"
                }
            }
        },
...
        "db.Booking": {
            "type": "object",
            "properties": {
                "booking_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.BookingItem"
                    }
                },
                "customer_id": {
                    "$ref": "#/definitions/db.User"
                },
                "event_id": {
                    "$ref": "#/definitions/db.Event"
                },
                "id": {
                    "type": "string"
                },
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Payment"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "db.BookingItem": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "$ref": "#/definitions/db.Booking"
                },
                "event_schedule_id": {
                    "$ref": "#/definitions/db.EventSchedule"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "qr": {
                    "type": "string"
                },
                "seat_id": {
                    "$ref": "#/definitions/db.Seat"
                },
                "status": {
                    "type": "string"
                },
                "ticket_id": {
                    "$ref": "#/definitions/db.Ticket"
                }
            }
        },
        "db.Category": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "db.Event": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Booking"
                    }
                },
                "category_id": {
                    "$ref": "#/definitions/db.Category"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "creator_id": {
                    "$ref": "#/definitions/db.User"
                },
                "description": {
                    "type": "string"
                },
                "event_schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.EventSchedule"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "preview_image": {
                    "type": "string"
                },
                "seat_zone": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.SeatZone"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Ticket"
                    }
                }
            }
        },
        "db.EventSchedule": {
            "type": "object",
            "properties": {
                "end_checkin_time": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start_checkin_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "db.Membership": {
            "type": "object",
            "properties": {
                "base_point": {
                    "type": "integer"
                },
                "discount": {
                    "description": "Directus will return a string even if they are set as decimal",
                    "type": "string"
                },
                "early_buy_time": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tier": {
                    "type": "string"
                }
            }
        },
        "db.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "booking_id": {
                    "$ref": "#/definitions/db.Booking"
                },
                "id": {
                    "type": "string"
                },
                "payment_gateway": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "refunds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Refund"
                    }
                },
                "status": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "db.Refund": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "payment_id": {
                    "$ref": "#/definitions/db.Payment"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "db.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "db.Seat": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "reserved_by": {
                    "$ref": "#/definitions/db.User"
                },
                "seat_number": {
                    "type": "string"
                },
                "seat_zone_id": {
                    "$ref": "#/definitions/db.SeatZone"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "db.SeatZone": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "event_id": {
                    "$ref": "#/definitions/db.Event"
                },
                "id": {
                    "type": "string"
                },
                "seats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Seat"
                    }
                },
                "status": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Ticket"
                    }
                },
                "total_seats": {
                    "type": "integer"
                }
            }
        },
        "db.Ticket": {
            "type": "object",
            "properties": {
                "base_price": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "event_id": {
                    "$ref": "#/definitions/db.Event"
                },
                "id": {
                    "type": "string"
                },
                "rank": {
                    "type": "string"
                },
                "seat_zone_id": {
                    "$ref": "#/definitions/db.SeatZone"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "db.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Booking"
                    }
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/db.Role"
                },
                "status": {
                    "type": "string"
                },
                "user_membership_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.UserMembershipLog"
                    }
                },
                "user_telegrams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.UserTelegram"
                    }
                }
            }
        },
        "db.UserMembershipLog": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "points_delta": {
                    "type": "integer"
                },
                "resulting_points": {
                    "type": "integer"
                }
            }
        },
        "db.UserTelegram": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "telegram_chat_id": {
                    "type": "string"
                },
                "user_id": {
                    "$ref": "#/definitions/db.User"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}